requirements.txt
cachetools
certifi
charset-normalizer
click
colorama
commonmark
et-xmlfile
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
grpcio
grpcio-status
heic-to-jpg
idna
Jinja2
loguru
lxml
MarkupSafe
mypy-extensions
numpy
openpyxl
pandas
pathlib
proto-plus
protobuf
pyasn1
pyasn1-modules
Pygments
python-dateutil
python-docx
python-dotenv
pytz
requests
rich
rsa
six
sortedcontainers
urllib3
win32-setctime
